requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comtypes==1.4.8
distro==1.9.0
docutils==0.21.2
exceptiongroup==1.2.2
google==3.0.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-texttospeech==2.21.1
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
openai==1.55.1
pdfminer==20191125
playsound==1.3.0
plyer==2.1.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.18.0
PyPDF2==3.0.1
pypiwin32==223
python-vlc==3.0.21203
pyttsx3==2.98
pywin32==308
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
